Change to centralized installation
The problem is that it is anyhow not useable with a read-only tools
volume, because it writes libraries to the installation directory.

We need to discuss how that works out inc ase different simulators
want to be used (also maybe in different versions). If there is no
simple workaround we probably need to change the procedure to first
copy/clone cocotb.